Issue #16: Square root not working.
This check-in fixes several bugs, improves logging during the square root step and improved the speed of the polynomial modulus a polynomial routine.

The square root method seems to succeed with a much higher probability than previously, but not with the probability of 1/2 as expected.

However, this might just be an issue with small number factorizations, in that its not picking an  inert prime that is large enough.
The formula for picking an inert prime starting value returns a proportionately sized inert prime. However, the algorithm seems to only succeed for small values of N when the inert prime is a significant fraction of the value of N.

More test values of a reasonably appropriate (larger) size are needed to avoid factoring by numerical coincidence and to avoid the inert prime too small issue mentioned above.

More testing needed.

Full list of changes:
- Changed target framework from 6.0 to 5.0 (until I can be bothered to upgrade my windows to version 10, )
- Added static logging function delegate: GNFS.LogFunction
- Added a bunch more logging around the square root step, as well as other places.
- Legendre.SymbolSearch had a bug in it. I fixed the bug.
- Polynomial.ModMod had a horrible runtime. I was able to write a version that ran in a fraction of the time by avoiding needless array allocations and diligently using modulus everywhere to avoid large numbers.
- Normal.Algebraic now calculates things using the arbitrary precision arithmetic library BigRational. This is important for large polynomials such as used when factoring RSA-100, as the norm calculations using only decimals would begin to lose accuracy and then eventually throw an overflow exception.
- There was a bug with SmoothRelations_TargetQuantity, in that it was often less than SmoothRelationsRequiredForMatrixStep, sometimes by 1, and so would needlessly run an extra round of sieving.
- I added some safeguards around the RecursiveDelete function in TestGNFS, moved it to a reusable helper class and made each test class call that common helper class function.
- Removed some dead code.